Added regex for vcpkg output
output now simplifies what vcpkg is throwing to stdout. still need to test on errors. still need to dump debug log in the case that vcpkg fails. included package counting which should make it easier in the future to circumvent the vcpkg lock when there is no extra work to be done by vcpkg