Added serial option for execution
Parallel runs with large number of presets can be buggy due to the fact that build systems were not designed to account for parallel execution. For example, MSBuild fails entirely if it can't open a file. --serial flag has been added to stop parallel execution while keeping the UI the same.